requirements.txt
langchain-pinecone
flask
-e .
streamlit
python-dotenv
langchain
langchain-pinecone
langchain-google-genai
pinecone-client
sentence-transformers
python-dotenv
langchain_experimental
pinecone[grpc]